refactor(Language): serialization for ComputeBase extended Components

+++ sources/Core/Application.h
@@ -138,6 +138,6 @@ namespace Nodable
 		/** Reflection using mirror*/
 		MIRROR_CLASS(Application)();
 
-        void closeFile(int _fileIndex);
+        void closeFile(size_t _fileIndex);
     };
 }
